+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.0] - 2021-07-06
+### Added
+ - Example Ruby scripts.
+ - Interactive Clojure interpreter (uses Clojure 1.10.3).
+ - Clojure script capability (uses Clojure 1.10.3).
+
+
 ## [0.1.0] - 2021-03-08
 ### Added
  - Interactive Ruby interpreter.